requirements.txt

python-docx>=0.8.11
PyPDF2>=2.12.1
Pillow>=9.5.0
requests>=2.28.0
